transcoder/requirements.txt
asyncio
paho-mqtt
python-dotenv
termcolor
attr
attrs
crcmod
nbformat
ruuvitag_sensor
bleak